Conflicts:
	public/javascripts/common/hive.FileUploader.js
	public/javascripts/service/hive.issue.View.js
	public/javascripts/service/hive.issue.Write.js

+++ app/controllers/PullRequestApp.java
@@ -10,9 +10,11 @@ import org.eclipse.jgit.api.errors.GitAPIException;
 import play.data.Form;
 import play.mvc.Controller;
 import play.mvc.Result;
+import play.mvc.Results;
 import playRepository.GitCommit;
 import playRepository.GitRepository;
 import playRepository.RepositoryService;
+import utils.Constants;
 import utils.JodaDateUtil;
 import views.html.git.create;
 import views.html.git.list;
@@ -42,13 +44,28 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result fork(String userName, String projectName) throws GitAPIException, IOException {
-        Project originalProject = Project.findByOwnerAndProjectName(userName, projectName);
-
         User currentUser = UserApp.currentUser();
+        if(currentUser.isAnonymous()) {
+            flash(Constants.WARNING, "user.login.alert");
+            return redirect(routes.UserApp.loginForm());
+        }
+
+        Project originalProject = Project.findByOwnerAndProjectName(userName, projectName);
+        if(originalProject == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
+
+        // private 프로젝트일 경우에는 해당 프로젝트의 멤버나 매니저만 fork할 수 있다.
+        if(!originalProject.isPublic) {
+            if(isGuest(originalProject, currentUser)) {
+                return badRequest("Only project's member and manager can fork private project");
+            }
+        }
 
         // 이미 포크한 프로젝트가 있다면 그 프로젝트로 이동.
         Project forkedProject = Project.findByOwnerAndOriginalProject(currentUser.loginId, originalProject);
         if(forkedProject != null) {
+            flash(Constants.WARNING, "already.existing.fork.alert");
             return redirect(routes.ProjectApp.project(forkedProject.owner, forkedProject.name));
         }
 
@@ -81,13 +98,11 @@ public class PullRequestApp extends Controller {
      * @throws ServletException
      */
     public static Result newPullRequestForm(String userName, String projectName) throws IOException, ServletException {
-        Project project = ProjectApp.getProject(userName, projectName);
-        if(project == null ) {
-            return notFound();
-        }
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
 
-        if(!project.isFork()) {
-            return badRequest("this request is allowed to only fork project");
+        Result result = validateBeforePullRequest(project, userName, projectName);
+        if(result != null) {
+            return result;
         }
 
         List<String> fromBranches = RepositoryService.getRepository(project).getBranches();
@@ -109,14 +124,16 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result newPullRequest(String userName, String projectName) {
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+
+        Result result = validateBeforePullRequest(project, userName, projectName);
+        if(result != null) {
+            return result;
+        }
+
         Form<PullRequest> form = new Form<>(PullRequest.class).bindFromRequest();
         if(form.hasErrors()) {
             return badRequest(form.errors().toString());
-        }
-
-        Project project = ProjectApp.getProject(userName, projectName);
-        if(!project.isFork()) {
-            return badRequest("this request is allowed to only fork project");
         }
 
         Project originalProject = project.originalProject;
@@ -146,7 +163,10 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result pullRequests(String userName, String projectName) {
-        Project project = ProjectApp.getProject(userName, projectName);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+        if(project == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
         List<PullRequest> pullRequests = PullRequest.findOpendPullRequests(project);
         return ok(list.render(project, pullRequests, "opened"));
     }
@@ -159,7 +179,10 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result closedPullRequests(String userName, String projectName) {
-        Project project = ProjectApp.getProject(userName, projectName);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+        if(project == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
         List<PullRequest> pullRequests = PullRequest.findClosedPullRequests(project);
         return ok(list.render(project, pullRequests, "closed"));
     }
@@ -172,7 +195,10 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result rejectedPullRequests(String userName, String projectName) {
-        Project project = ProjectApp.getProject(userName, projectName);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+        if(project == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
         List<PullRequest> pullRequests = PullRequest.findRejectedPullRequests(project);
         return ok(list.render(project, pullRequests, "rejected"));
     }
@@ -185,7 +211,10 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result sentPullRequests(String userName, String projectName) {
-        Project project = ProjectApp.getProject(userName, projectName);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+        if(project == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
         List<PullRequest> pullRequests = PullRequest.findSentPullRequests(project);
         return ok(list.render(project, pullRequests, "sent"));
     }
@@ -201,10 +230,12 @@ public class PullRequestApp extends Controller {
      * @return
      */
     public static Result pullRequest(String userName, String projectName, long pullRequestId) {
-        Project project = ProjectApp.getProject(userName, projectName);
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
-        if(pullRequest == null) {
-            return notFound();
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+
+        Result result = validatePullRequest(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
         }
 
         boolean isSafe = false;
@@ -215,12 +246,24 @@ public class PullRequestApp extends Controller {
         return ok(view.render(project, pullRequest, isSafe));
     }
 
+    /**
+     * {@code userName}과 {@code projectName}에 해당하는 프로젝트로 들어온
+     * {@code pullRequestId}에 해당하는 코드 요청의 커밋 목록을 조회한다.
+     *
+     * @param userName
+     * @param projectName
+     * @param pullRequestId
+     * @return
+     */
     public static Result pullRequestCommits(String userName, String projectName, long pullRequestId) {
-        Project project = ProjectApp.getProject(userName, projectName);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
-        if(pullRequest == null) {
-            return notFound();
+
+        Result result = validatePullRequest(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
         }
+
         List<GitCommit> commits = GitRepository.getPullingCommits(pullRequest);
         return ok(viewCommits.render(project, pullRequest, commits));
     }
@@ -237,6 +280,13 @@ public class PullRequestApp extends Controller {
      */
     public static Result accept(String userName, String projectName, long pullRequestId) {
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+
+        Result result = validatePullRequestOperation(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
+        }
+
         GitRepository.merge(pullRequest);
         if(pullRequest.state == State.CLOSED) {
             pullRequest.received = JodaDateUtil.now();
@@ -245,6 +295,8 @@ public class PullRequestApp extends Controller {
         }
         return redirect(routes.PullRequestApp.pullRequest(userName, projectName, pullRequestId));
     }
+
+
 
     /**
      * {@code pullRequestId}에 해당하는 코드 요청을 보류한다.
@@ -258,6 +310,13 @@ public class PullRequestApp extends Controller {
      */
     public static Result reject(String userName, String projectName, Long pullRequestId) {
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+
+        Result result = validatePullRequestOperation(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
+        }
+
         pullRequest.state = State.REJECTED;
         pullRequest.received = JodaDateUtil.now();
         pullRequest.receiver = UserApp.currentUser();
@@ -277,6 +336,13 @@ public class PullRequestApp extends Controller {
      */
     public static Result open(String userName, String projectName, Long pullRequestId) {
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+
+        Result result = validatePullRequestOperation(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
+        }
+
         pullRequest.state = State.OPEN;
         pullRequest.received = JodaDateUtil.now();
         pullRequest.receiver = UserApp.currentUser();
@@ -294,6 +360,21 @@ public class PullRequestApp extends Controller {
      */
     public static Result cancel(String userName, String projectName, Long pullRequestId) {
         PullRequest pullRequest = PullRequest.findById(pullRequestId);
+        Project project = Project.findByOwnerAndProjectName(userName, projectName);
+        User user = UserApp.currentUser();
+
+        Result result = validatePullRequest(project, pullRequest, userName, projectName, pullRequestId);
+        if(result != null) {
+            return result;
+        }
+
+        // 게스트 중에서 코드 요청을 보낸 사용자는 취소 할 수 있다.
+        if(isGuest(project, user)) {
+            if(!user.equals(pullRequest.contributor)) {
+                forbidden("Only this project's member and manager and the pull_request's author are allowed.");
+            }
+        }
+
         pullRequest.delete();
         return redirect(routes.PullRequestApp.pullRequests(userName, projectName));
     }
@@ -321,6 +402,109 @@ public class PullRequestApp extends Controller {
                 return newProjectName;
             }
         }
+    }
+
+
+    /**
+     * {@link PullRequest}를 만들어 보내기 전에 프로젝트와 현재 사용자의 권한을 검증한다.
+     *
+     * when: {@link PullRequest}를 생성하는 폼이나 폼 서브밋을 처리하기 전에 사용한다.
+     *
+     * 현재 사용자가 익명 사용자일 경우 로그인 페이지로 이동하는 303 응답을 반환한다.
+     * 프로젝트가 null이면 400(bad request) 응답을 반환한다.
+     * 프로젝트가 Fork 프로젝트가 아닐 경우에도 400(bad request) 응답을 반환한다.
+     * 사용자의 프로젝트 권한이 guest일 경우에도 400(bad request) 응답을 반환한다.
+     * 이외의 경우에는 null을 반환한다.
+     *
+     * @param userName
+     * @param projectName
+     * @return
+     */
+    private static Result validateBeforePullRequest(Project project, String userName, String projectName) {
+        User currentUser = UserApp.currentUser();
+        if(currentUser.isAnonymous()) {
+            flash(Constants.WARNING, "user.login.alert");
+            return redirect(routes.UserApp.loginForm());
+        }
+
+        if(project == null ) {
+            return badRequestForNullProject(userName, projectName);
+        }
+        if(!project.isFork()) {
+            return badRequest("Only fork project is allowed this request");
+        }
+
+        // anonymous는 위에서 걸렀고, 남은건 manager, member, site-manager, guest인데 이중에서 guest만 다시 걸러낸다.
+        if(isGuest(project, currentUser)) {
+            return badRequest("Guest is not allowed this request");
+        }
+
+        return null;
+    }
+
+    private static boolean isGuest(Project project, User currentUser) {
+        return ProjectUser.roleOf(currentUser.loginId, project).equals("guest");
+    }
+
+    /**
+     * {@code userName}과 {@code projectName}에 해당하는 프로젝트가 없을 경우에 보내는 응답
+     *
+     * @param userName
+     * @param projectName
+     * @return
+     */
+    private static Status badRequestForNullProject(String userName, String projectName) {
+        return badRequest("No project matches given parameters'" + userName + "' and project_name '" + projectName + "'");
+    }
+
+    /**
+     * {@code project}과 {@code pullRequest}가 null인지 확인한다.
+     *
+     * @param project
+     * @param pullRequest
+     * @param userName
+     * @param projectName
+     * @param pullRequestId
+     * @return
+     */
+    private static Result validatePullRequest(Project project, PullRequest pullRequest,
+                                              String userName, String projectName, long pullRequestId) {
+        if(project == null) {
+            return badRequestForNullProject(userName, projectName);
+        }
+
+        if(pullRequest == null) {
+            return badRequest("No pull_request matches given pull_request_id '" + pullRequestId + "'");
+        }
+        return null;
+    }
+
+    /**
+     * {@code project}와 {@code pullRequest}가 null인지 확인하고
+     * 현재 사용자가 {@code project}의 코드 요청을 처리 할 수 있는지 확인한다.
+     *
+     * @param project
+     * @param pullRequest
+     * @param userName
+     * @param projectName
+     * @param pullRequestId
+     * @return
+     */
+    private static Result validatePullRequestOperation(Project project, PullRequest pullRequest,
+                                                       String userName, String projectName, long pullRequestId) {
+        User user = UserApp.currentUser();
+        if(user.isAnonymous()) {
+            flash(Constants.WARNING, "user.login.alert");
+            return redirect(routes.UserApp.loginForm());
+        }
+
+        Result result = validatePullRequest(project, pullRequest, userName, projectName, pullRequestId);
+
+        if(isGuest(project, user)) {
+            result = forbidden("Guest is not allowed this request");
+        }
+
+        return result;
     }
 
 }
